% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brsmatching.R
\name{balance_columns}
\alias{balance_columns}
\title{Compute balance covariate indicators}
\usage{
balance_columns(
  df,
  id = "id",
  time = "time",
  trt_time = "trt_time",
  balance_covariates = NULL
)
}
\arguments{
\item{df}{data frame containing columns matching the \code{id, time,
  trt_time} arguments, and covariates. This data frame is expected to be in
tidy, long format, so that the \code{id}, \code{trt_time}, and baseline
variables may be repeated for different values of \code{time}.  Data frame
should be unique at \code{id} and \code{time}.}

\item{id}{optional parameter to specify the name of the id column in
\code{df}.}

\item{time}{optional parameter to specify the name of the time column in
\code{df}.}

\item{trt_time}{optional parameter to specify the name of the treatment time
column in \code{df}.}

\item{balance_covariates}{optional parameter to specify the names of
covariates to be used in balancing. If \code{NULL}, will default to all
columns except those named by \code{id, time, trt_time}.}
}
\value{
a matrix with the same number of rows as the input dataframe.  Each
column is an indicator variable corresponds to a balancing criteria.  The
number of balance columns returned is 2*n_numeric_cols + n_factor_cols.
}
\description{
The \code{balance_columns} function takes longitudinal data input and returns
corresponding balance columns on desired covariates according to the process
described in Li et al. (2001) "Balanced Risk Set Matching". Each balance
column is an indicator variable.
}
\details{
For numeric covariates, quantiles are computed for treated individuals at
their treated time, and indicator variables for whether each covariate
exceeds the quantiles are returned.  Default quantiles used are 1/3 and 2/3.

For factor and character covariates, indicators for each level (except one)
are returned as would be done in a call to \code{stats::model.matrix}.
}
\examples{
df <- data.frame(
   hhidpn = rep(1:3, each = 3),
   wave = rep(1:3, 3),
  treatment_time = rep(c(2,3,NA), each = 3),
  X1 = c(2,2,2,3,3,3,9,9,9),
  X2 = rep(c("a","a","b"), each = 3),
  X3 = c(9,4,5,6,7,2,3,4,8),
  X4 = c(8,9,4,5,6,7,2,3,4)
)

balance_covariates <- c("X1", "X2", "X3", "X4")
bal <- balance_columns(df, "hhidpn", "wave", "treatment_time",
                       balance_covariates = balance_covariates)

}
