% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brsmatching.R
\name{brsmatch}
\alias{brsmatch}
\title{Balanced Risk Set Matching}
\usage{
brsmatch(
  n_pairs,
  df,
  id = "id",
  time = "time",
  trt_time = "trt_time",
  covariates = NULL,
  balance_covariates = NULL,
  optimizer = "gurobi",
  verbose = FALSE,
  balance = TRUE
)
}
\arguments{
\item{n_pairs}{number of pairs desired from matching}

\item{df}{data frame containing columns matching the \code{id, time, trt_time} arguments, and covariates.
This data frame is expected to be in tidy, long format, so that the \code{id}, \code{trt_time}, and
baseline  variables may be repeated for different values of \code{time}.  Data frame should be unique
at \code{id} and \code{time}.}

\item{id}{optional parameter to specify the name of the id column in \code{df}.}

\item{time}{optional parameter to specify the name of the time column in \code{df}.}

\item{trt_time}{optional parameter to specify the name of the treatment time column in \code{df}.}

\item{covariates}{optional parameter to specify the names of covariates to be used. If \code{NULL}, will
default to all columns except those named by \code{id, time, trt_time}.}

\item{balance_covariates}{optional parameter to specify the names of covariates to be used in balancing.
If \code{NULL}, will default to all columns except those named by \code{id, time, trt_time}.}

\item{optimizer}{"gurobi" or "glpk". Specifies which optimizer output to use;
defaults to "gurobi".}

\item{verbose}{logical; if TRUE, will print some useful information for
potentially long model calls; defaults to FALSE}

\item{balance}{When TRUE, a balanced risk set matching model will be built.
When FALSE, or when bal_all = NULL, balancing constraints will not be
included.}
}
\value{
TODO:
}
\description{
TODO: Come up with a better description here.  Mention the Li et al. (2001) Balanced Risk Set Matching
paper.
}
\examples{



}
