% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmatch.R
\name{coxph_match1}
\alias{coxph_match1}
\title{Propensity Score Matching with Time-Dependent Covariates}
\usage{
coxph_match1(
  df,
  id = "id",
  time = "time",
  trt_time = "trt_time",
  covariates = NULL
)
}
\arguments{
\item{df}{data frame containing columns matching the \code{id, time, trt_time} arguments, and covariates.
This data frame is expected to be in tidy, long format, so that the \code{id}, \code{trt_time}, and
baseline  variables may be repeated for different values of \code{time}.  Data frame should be unique
at \code{id} and \code{time}.}

\item{id}{optional parameter to specify the name of the id column in \code{df}.}

\item{time}{optional parameter to specify the name of the time column in \code{df}.}

\item{trt_time}{optional parameter to specify the name of the treatment time column in \code{df}.}

\item{covariates}{optional parameter to specify the names of covariates to be used. If \code{NULL}, will
default to all columns except those named by \code{id, time, trt_time}.}
}
\value{
a data frame with matched pairs based on propensity scores generated
using a Cox proportional hazards model. This data frame will have three
columns.  \code{trt.id} refers to the treated id, \code{con.id} refers to
the id that is a possible control, \code{distance} refers to the scaled
euclidean distance generated based on the propensity scores.
}
\description{
Perform propensity score matching as described in Lu (2005) "Propensity Score
Matching with Time-Dependent Covariates".  Given a longitudinal data frame
with covariate information, along with treatment time, build a MIP problem
that matches treated individuals to those that haven't been treated yet (or
are never treated) based on time-dependent propensity scores from a Cox
proportional hazards model. If balancing is desired, the model will try to
minimize the imbalance in terms of specified balancing covariates in the
final pair output.  Each treated individual is matched to one other
individual.
}
\examples{
df <- data.frame(
  hhidpn = rep(1:3, each = 3),
  wave = rep(1:3, 3),
  treatment_time = rep(c(2,3,NA), each = 3),
  X1 = c(2,2,2,3,3,3,9,9,9),
  X2 = rep(c("a","a","b"), each = 3),
  X3 = c(9,4,5,6,7,2,3,4,8),
  X4 = c(8,9,4,5,6,7,2,3,4)
)

coxph_match1(df = df, id = "hhidpn", time = "wave", trt_time = "treatment_time")

}
