% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brsmatch.R
\name{output_pairs}
\alias{output_pairs}
\title{Output pairs to new format}
\usage{
output_pairs(matched_ids, id = "id", id_list = NULL)
}
\arguments{
\item{matched_ids}{data frame with two columns: trt_id, all_id.  Each row
consists of a matched pair, where "trt_id" provides the id of the treated
and "all_id" provides the id of the control.}

\item{id}{optional parameter to specify the name of the id column in
\code{df}.}

\item{id_list}{optional vector of ids to include in the output.}
}
\value{
a data frame with columns "id", "pair_id", and "type".  "id" refers
to the individual ids, "pair_id" is a unique identifier for each pair, and
"type" indicates whether the id is from treatment ("trt") or control ("all")
}
\description{
Takes a data frame with each row as a pair and returns output in long, tidy
format that indicates the matched pairs
}
\examples{
matched_ids <- data.frame(
  trt_id = c(2, 5, 7),
  all_id = c(3, 1, 4)
)

output_pairs(matched_ids)

}
