% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brsmatching.R
\name{balance_columns}
\alias{balance_columns}
\title{Compute balance covariate indicators}
\usage{
balance_columns(
  df,
  id = "id",
  time = "time",
  trt_time = "trt_time",
  balance_covariates = NULL
)
}
\arguments{
\item{df}{data frame containing columns matching the \code{id, time, trt_time} arguments, and covariates.
This data frame is expected to be in tidy, long format, so that the \code{id}, \code{trt_time}, and
baseline  variables may be repeated for different values of \code{time}.  Data frame should be unique
at \code{id} and \code{time}.}

\item{id}{optional parameter to specify the name of the id column in \code{df}.}

\item{time}{optional parameter to specify the name of the time column in \code{df}.}

\item{trt_time}{optional parameter to specify the name of the treatment time column in \code{df}.}

\item{balance_covariates}{optional parameter to specify the names of covariates to be used in balancing.
If \code{NULL}, will default to all columns except those named by \code{id, time, trt_time}.}
}
\value{
TODO: add return description
}
\description{
TODO: add description
}
\examples{
df <- data.frame(
  treatment_time = rep(c(2,3,NA), each = 3),
  X1 = c(2,2,2,3,3,3,9,9,9),
  X2 = rep(c("a","a","b"), each = 3),
  X3 = c(9,4,5,6,7,2,3,4,8),
  X4 = c(8,9,4,5,6,7,2,3,4)
)

balance_covariates <- c("X1", "X2", "X3", "X4")
bal <- balance_columns(df, "hhidpn", "wave", "treatment_time", balance_covariates = balance_covariates)

}
