% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brsmatch.R
\name{brsmatch}
\alias{brsmatch}
\title{Balanced Risk Set Matching}
\usage{
brsmatch(
  n_pairs,
  data,
  id = "id",
  time = "time",
  trt_time = "trt_time",
  covariates = NULL,
  balance_covariates = NULL,
  optimizer = c("glpk", "gurobi"),
  verbose = FALSE,
  balance = TRUE,
  options = c("none", "between period treatment")
)
}
\arguments{
\item{n_pairs}{The number of pairs desired from matching.}

\item{data}{A data.frame or similar containing columns matching the \verb{id, time, trt_time} arguments, and covariates. This data frame is expected to
be in tidy, long format, so that \code{id}, \code{trt_time}, and other variables may
be repeated for different values of \code{time}. The data.frame should be unique
at \code{id} and \code{time}.}

\item{id}{A character specifying the id column name (default \code{'id'}).}

\item{time}{A character specifying the time column name (default \code{'time'}).}

\item{trt_time}{A character specifying the treatment time column name
(default \code{'trt_time'}).}

\item{covariates}{A character vector specifying the covariates to use for
matching (default \code{NULL}). If \code{NULL}, this will default to all columns
except those named by the \code{id}, \code{time}, and \code{trt_time} arguments.}

\item{balance_covariates}{A character vector specifying the covariates to use
for balancing (default \code{NULL}).  If \code{NULL}, this will default to all
columns except those named by the \code{id}, \code{time}, and \code{trt_time} arguments.}

\item{optimizer}{The optimizer to use (default \code{'glpk'}). The option
\code{'gurobi'} requires an external license and package, but offers speed
improvements.}

\item{verbose}{A logical value indicating whether to print information to the
console during a potentially long matching process.}

\item{balance}{A logical value indicating whether to including balancing
constraints in the matching process.}

\item{options}{Some additional arguments for rare scenarios (default
\code{'none'}).  If options = 'between period treatment' then matches must be
made on the time period preceding treatment and some special care must be
taken.}
}
\value{
A data frame containing the pair information.  The data frame has
columns \code{id}, \code{pair_id}, and \code{type}. \code{id} matches the input parameter and
will contain all ids from the input data frame.  \code{pair_id} refers to the id
of the computed pairs; \code{NA} values indicate unmatched individuals.  \code{type}
indicates whether the individual in the pair is considered as treatment
("trt") or control ("all") in that pair.
}
\description{
Perform balanced risk set matching as described in Li et al. (2001) "Balanced
Risk Set Matching".  Given a longitudinal data frame with covariate
information, along with treatment time, build a MIP problem that matches
treated individuals to those that haven't been treated yet (or are never
treated) based on minimizing the Mahalanobis distance between covariates. If
balancing is desired, the model will try to minimize the imbalance in terms
of specified balancing covariates in the final pair output.  Each treated
individual is matched to one other individual.
}
\examples{
df <- data.frame(
  hhidpn = rep(1:3, each = 3),
  wave = rep(1:3, 3),
  treatment_time = rep(c(2,3,NA), each = 3),
  X1 = c(2,2,2,3,3,3,9,9,9),
  X2 = rep(c("a","a","b"), each = 3),
  X3 = c(9,4,5,6,7,2,3,4,8),
  X4 = c(8,9,4,5,6,7,2,3,4)
)

if (requireNamespace("Rglpk", quietly = TRUE)) {
  brsmatch(n_pairs = 1, data = df, id = "hhidpn", time = "wave",
           trt_time = "treatment_time", optimizer = "glpk")
}

}
\author{
Sean Kent
}
