% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmatch.R
\name{coxpsmatch}
\alias{coxpsmatch}
\title{Propensity Score Matching with Time-Dependent Covariates}
\usage{
coxpsmatch(
  n_pairs = 10^10,
  df,
  id = "id",
  time = "time",
  trt_time = "trt_time",
  covariates = NULL,
  exact_match = NULL
)
}
\arguments{
\item{n_pairs}{number of pairs desired from matching}

\item{df}{data frame containing columns matching the \code{id, time,
  trt_time} arguments, and covariates. This data frame is expected to be in
tidy, long format, so that the \code{id}, \code{trt_time}, and baseline
variables may be repeated for different values of \code{time}.  Data frame
should be unique at \code{id} and \code{time}.}

\item{id}{optional parameter to specify the name of the id column in
\code{df}.}

\item{time}{optional parameter to specify the name of the time column in
\code{df}.}

\item{trt_time}{optional parameter to specify the name of the treatment time
column in \code{df}.}

\item{covariates}{optional parameter to specify the names of covariates to be
used. If \code{NULL}, will default to all columns except those named by
\code{id, time, trt_time}.}

\item{exact_match}{A vector of optional covariates to perform exact matching
on. If \code{NULL}, no exact matching is done.}
}
\value{
a data frame containing the pair information.  The data frame has
columns \code{id}, "pair_id", and "type". \code{id} matches the input
parameter and will contain all ids from the input data frame.  "pair_id"
refers to the id of the computed pairs; NA values indicate unmatched
individuals.  "type" indicates whether the individual in the pair is
considered as treatment ("trt") or control ("all") in that pair.
}
\description{
Perform propensity score matching as described in Lu (2005) "Propensity Score
Matching with Time-Dependent Covariates".  Given a longitudinal data frame
with covariate information, along with treatment time, build a MIP problem
that matches treated individuals to those that haven't been treated yet (or
are never treated) based on time-dependent propensity scores from a Cox
proportional hazards model. If balancing is desired, the model will try to
minimize the imbalance in terms of specified balancing covariates in the
final pair output.  Each treated individual is matched to one other
individual.
}
\examples{
df <- data.frame(
  hhidpn = rep(1:3, each = 3),
  wave = rep(1:3, 3),
  treatment_time = rep(c(2,3,NA), each = 3),
  X1 = c(2,2,2,3,3,3,9,9,9),
  X2 = rep(c("a","a","b"), each = 3),
  X3 = c(9,4,5,6,7,2,3,4,8),
  X4 = c(8,9,4,5,6,7,2,3,4)
)

coxpsmatch(n_pairs = 1, df = df, id = "hhidpn", time = "wave", trt_time = "treatment_time")

}
\author{
Mitchell Paukner
}
